requirements.txt
click==8.1.2
cryptography==3.4.7
discord.py==1.7.3
google_api_python_client==2.46.0
httplib2==0.20.4
matplotlib==3.3.4
oauth2client==4.1.3
requests==2.26.0
scapy==2.4.5
stegano==0.10.2
